Refactor API to controller
