+++ src/main/java/edu/gcc/controller/SearchAPI.java
@@ -1,8 +1,7 @@
-package edu.gcc.api;
+package edu.gcc.controller;
 
 import edu.gcc.Course;
 import edu.gcc.Search;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
